Update code to live in local system bin
Warning: The service is still being run as a user and *NOT* as system,
even though the notify-listener is being placed in system directories.
Root would be able to change the script (or anyone else with sudo
powers) and be able to compromise a user's session. The script should be
chown'd to root:root and chmod'd to 0655.